rename TallCard and update imports
