+++ src/components/Cards/TallCard/TallCard.tsx
